Added support for linux shared mode
Added support for "x64-linux-shared"

Added Linux-Shared support to port "openssl-linux".

Added Linux-Shared support for port "libvorbis"

Fixed build issues with ffmpeg, liblzma, libwebsockets and zstd.

Fixed compilation issues on Windows.

Fixed issues on Windows.

Fixed issue with "pixman" port on Linux.

Fixed build issue with "pango" port on Linux.

Fixed issues on Windows.

[TASK-029]: Fixed build issue on CentOS 7.

Glib: Fixed build issues with dtrace and libselinux (CentOS).

Fixed compilation issue with LIBSELINUX on Windows.

Fixed issue with wrong pointer size in fontconfig for target x64-linux-shared.

Fixed build issue on Linux CentOS.

[TASK-040]: added System-Hook to Boost libraries

Fixed build issues after rebase with master.

[TASK-040]: Fixed build issues on Linux CentOS.

+++ ports/ffmpeg/portfile.cmake
@@ -365,11 +365,15 @@ endif()
 
 set(OPTIONS_DEBUG "--debug") # Note: --disable-optimizations can't be used due to http://ffmpeg.org/pipermail/libav-user/2013-March/003945.html
 set(OPTIONS_RELEASE "")
+set(OPTIONS_DEBUG "--enable-debug")
 
 set(OPTIONS "${OPTIONS} ${OPTIONS_CROSS}")
+set(OPTIONS "${OPTIONS} --extra-cflags=-DHAVE_UNISTD_H=0")
+set(OPTIONS "${OPTIONS} --arch=${VCPKG_TARGET_ARCHITECTURE}")
 
 if(VCPKG_LIBRARY_LINKAGE STREQUAL "dynamic")
     set(OPTIONS "${OPTIONS} --disable-static --enable-shared")
+	set(OPTIONS "${OPTIONS} --enable-pic")
     if (VCPKG_TARGET_IS_UWP)
         set(OPTIONS "${OPTIONS} --extra-ldflags=-APPCONTAINER --extra-ldflags=WindowsApp.lib")
     endif()
@@ -487,7 +491,9 @@ if("ffplay" IN_LIST FEATURES)
     vcpkg_copy_tools(TOOL_NAMES ffplay AUTO_CLEAN)
 endif()
 
+if(NOT DEFINED VCPKG_BUILD_TYPE OR VCPKG_BUILD_TYPE STREQUAL "debug")
 file(REMOVE_RECURSE ${CURRENT_PACKAGES_DIR}/debug/include ${CURRENT_PACKAGES_DIR}/debug/share)
+endif()
 
 if(VCPKG_LIBRARY_LINKAGE STREQUAL "static")
     file(REMOVE_RECURSE ${CURRENT_PACKAGES_DIR}/bin ${CURRENT_PACKAGES_DIR}/debug/bin)
